pyproject.toml
[project]
name = "dify-runtime-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "openai>=2.8.1",
    "psycopg2-binary>=2.9.11",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.44",
]
